MVP: add work command compatibility for repo workbench
Adds prometheos work ... compatibility for the Repo Workbench MVP.
The Repo Workbench golden path can now be accessed through the
product-facing work command surface while keeping the existing
repo command available.

- Extracted reusable service functions into src/repo_workbench.rs
- Added work create --repo ... delegation
- Work run, artifacts, continue, memory show detect Repo Workbench
  context IDs via file-backed store check
- Added work approve <artifact_id> for approval recording
- Both repo and work commands share same implementation
- Updated docs to show product-facing work commands
- Tests cover delegation and context detection
